patch save e delete preferiti per utente loggato, delete utente loggato e test api

+++ src/main/java/it/epicode/endgame/controller/UtenteController.java
@@ -5,8 +5,11 @@ import it.epicode.endgame.dto.UpdateUtenteRequest;
 import it.epicode.endgame.dto.UpdateVideogiocoRequest;
 import it.epicode.endgame.dto.UtenteRequest;
 import it.epicode.endgame.exception.BadRequestException;
+import it.epicode.endgame.exception.UnAuthorizedException;
+import it.epicode.endgame.model.Tipologia;
 import it.epicode.endgame.model.Utente;
 import it.epicode.endgame.model.Videogioco;
+import it.epicode.endgame.security.JwtTools;
 import it.epicode.endgame.service.UtenteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -26,6 +29,8 @@ public class UtenteController {
     private UtenteService utenteService;
     @Autowired
     private Cloudinary cloudinary;
+    @Autowired
+    private JwtTools jwtTools;
     @GetMapping("/utenti")
     public List<Utente> getAll(){
         return utenteService.getAllUtenti();
@@ -37,13 +42,24 @@ public class UtenteController {
     }
 
     @PutMapping("/utenti/{id}")
-    public Utente updateUtente(@PathVariable int id, @RequestBody @Validated UtenteRequest utenteRequest, BindingResult bindingResult){
+    public Utente updateUtente(@PathVariable int id, @RequestHeader("Authorization") String token, @RequestBody @Validated UtenteRequest utenteRequest, BindingResult bindingResult){
         if(bindingResult.hasErrors()){
             throw new BadRequestException(bindingResult.getAllErrors().toString());
         }
 
-        return utenteService.updateUtente(id, utenteRequest);
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new UnAuthorizedException("Token non fornito o non valido");
+        }
 
+        String jwt = token.substring(7);
+        String userIdFromToken = jwtTools.extractUserIdFromToken(jwt);
+
+        Utente utente = utenteService.getUtenteById(id);
+        if (utente.getIdUtente() != Integer.parseInt(userIdFromToken)) {
+            throw new UnAuthorizedException("L'utente non è autorizzato a modificare questo profilo");
+        }
+
+        return utenteService.updateUtente(id, utenteRequest);
     }
     @PostMapping("/utenti")
     public Utente save(@RequestBody @Validated UtenteRequest utenteRequest, BindingResult bindingResult){
@@ -54,14 +70,39 @@ public class UtenteController {
         return utenteService.saveUtente(utenteRequest);
     }
     @DeleteMapping("/utenti/{id}")
-    public void deleteUtente(@PathVariable int id){
+    public String deleteUtente(@PathVariable int id, @RequestHeader("Authorization") String token) {
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new UnAuthorizedException("Token non fornito o non valido");
+        }
+
+        String jwt = token.substring(7);
+        String userIdFromToken = jwtTools.extractUserIdFromToken(jwt);
+
+        Utente utente = utenteService.getUtenteById(id);
+//        if (!utente.getTipologia().equals(Tipologia.ADMIN))
+            if (utente.getIdUtente() != Integer.parseInt(userIdFromToken)) {
+            throw new UnAuthorizedException("L'utente non è autorizzato a eliminare questo profilo");
+        }
 
         utenteService.deleteUtente(id);
+        return "Utente cancellato";
     }
     @PatchMapping("/utenti/{id}")
-    public Utente updateUtentePatch(@PathVariable int id, @RequestBody @Validated UpdateUtenteRequest updateUtenteRequest, BindingResult bindingResult) {
+    public Utente updateUtentePatch(@PathVariable int id, @RequestHeader("Authorization") String token, @RequestBody @Validated UpdateUtenteRequest updateUtenteRequest, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             throw new BadRequestException(bindingResult.getAllErrors().toString());
+        }
+
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new UnAuthorizedException("Token non fornito o non valido");
+        }
+
+        String jwt = token.substring(7);
+        String userIdFromToken = jwtTools.extractUserIdFromToken(jwt);
+
+        Utente utente = utenteService.getUtenteById(id);
+        if (utente.getIdUtente() != Integer.parseInt(userIdFromToken)) {
+            throw new UnAuthorizedException("L'utente non è autorizzato a modificare questo profilo");
         }
         return utenteService.updateUtentePatch(id, updateUtenteRequest);
     }
@@ -70,17 +111,52 @@ public class UtenteController {
         return utenteService.updateTipologiaUtente(id, tipologia);
     }
     @PatchMapping("/utenti/{idUtente}/preferiti/{idVideogioco}")
-    public Utente aggiungiPreferito(@PathVariable int idUtente, @PathVariable int idVideogioco) {
+    public Utente aggiungiPreferito(@PathVariable int idUtente, @PathVariable int idVideogioco, @RequestHeader("Authorization") String token) {
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new UnAuthorizedException("Token non fornito o non valido");
+        }
+
+        String jwt = token.substring(7);
+        String userIdFromToken = jwtTools.extractUserIdFromToken(jwt);
+
+        if (idUtente != Integer.parseInt(userIdFromToken)) {
+            throw new UnAuthorizedException("L'utente non è autorizzato ad aggiungere preferiti per questo profilo");
+        }
+
         return utenteService.savePreferitiUtente(idUtente, idVideogioco);
     }
     @GetMapping("/utenti/{idUtente}/preferiti")
-    public Page<Videogioco> getPreferitiPaginati(@PathVariable int idUtente, Pageable pageable) {
+    public Page<Videogioco> getPreferitiPaginati(@PathVariable int idUtente, Pageable pageable, @RequestHeader("Authorization") String token) {
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new UnAuthorizedException("Token non fornito o non valido");
+        }
+
+        String jwt = token.substring(7);
+        String userIdFromToken = jwtTools.extractUserIdFromToken(jwt);
+
+        if (idUtente != Integer.parseInt(userIdFromToken)) {
+            throw new UnAuthorizedException("L'utente non è autorizzato a visualizzare i preferiti per questo profilo");
+        }
+
         return utenteService.findPaginatedPreferitiUtente(idUtente, pageable);
     }
+
     @DeleteMapping("/utenti/{idUtente}/preferiti/{idVideogioco}")
-    public Utente rimuoviPreferito(@PathVariable int idUtente, @PathVariable int idVideogioco) {
+    public Utente rimuoviPreferito(@PathVariable int idUtente, @PathVariable int idVideogioco, @RequestHeader("Authorization") String token) {
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new UnAuthorizedException("Token non fornito o non valido");
+        }
+
+        String jwt = token.substring(7);
+        String userIdFromToken = jwtTools.extractUserIdFromToken(jwt);
+
+        if (idUtente != Integer.parseInt(userIdFromToken)) {
+            throw new UnAuthorizedException("L'utente non è autorizzato a rimuovere preferiti per questo profilo");
+        }
+
         return utenteService.removePreferitiUtente(idUtente, idVideogioco);
     }
+
     @PatchMapping("/utenti/{id}/upload")
     public Utente uploadAvatar(@PathVariable int id, @RequestParam("upload") MultipartFile file) throws IOException {
         return utenteService.uploadAvatar(id,
